+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 - Added: Generator command for a ViewModel class
 - Added: Generator command for data patches
+- Added: Generator command for cron jobs
 - Added: Jump-to-definition for magento modules (in module.xml and routes.xml)
 - Fixed: Method plugin hover messages are now grouped and include a link to di.xml